don't log.info to stderr
Summary:
We currently emit all Hh_logger.log messages to stderr.

(1) Everything from stderr is recorded in the LSP transcripts by VSCode. For folks in other teams who read the transcripts, they see all these "STDERR" messages and don't know what they mean, and they think there are bugs in Hack.

(2) We currently Hh_logger.log every single file-changed message. For folks with several thousand file-changes, the sheer volume of sending all these log messages over stderr is enough to saturate a low-bandwidth connection. End result is that VSCode is unusable for up to a minute.

This diff fixes both. It makes it so stderr in the client only gets LogLevel.Error messages, while file continues to get LogLevel.Info messages.

We have a pre-existing LSP flag "--verbose" which is used notably in debugging and tests, to get a full transcript to stderr. The documentation for this flag says "verbose logs to stderr and `hh --ide-logname` and `--lsp-logname`"

In this diff I have preserved that behavior: the --verbose flag will put LogLevel.Debug on both stderr and file.

Reviewed By: arxanas

Differential Revision: D19530887

fbshipit-source-id: cd4572fec70aa72d5005b8e6ceb4b7b769eeb853