show useful message 1
Summary:
End goal of this stack is to surface to the user the useful actionable error messages that come from external tools.

This diff restructures Saved_state_loader.load_error. Previously it was basically only fit for log messages and telemetry.

This diff restructures it to have a user-facing message separate from a log-facing string. And this diff also audits each of the places that produce a load_error, to make sure they're creating the right information.

Reviewed By: arxanas

Differential Revision: D19529849

fbshipit-source-id: e0e7ac51f393996ea060364eb10bfbf783404af6