show useful messages 2
Summary:
End goal is to surface to the user actionable information.

At the moment, the interface between clientlsp.ml and ClientIdeService.initialize, in case of error, is a Marshal_tools.remote_exception_data which includes a string (suitable for user-facing message) and a callstack (suitable for logging).

But we'll want more. We'll also want to expose whether or not the user-facing message demands action from the user, or merely represents some kind of bug on our side.

Therefore, this diff changes the interface to be our own structure, ClientIdeMessage.edata (short for "error_data"). At the moment it still only contains string+callstack. We'll add is_actionable to it in the next diff.

Reviewed By: arxanas

Differential Revision: D19529854

fbshipit-source-id: 5c08bf1113218bebbd8ccc81debf868dc255597b