Non-generic api to parser errors
Summary:
Generic API makes all the cost of this crate be shifted into `aast_parser`, where the compilation of it is serialized with other expensive things for no reason.

While there are two existing instantiations, they differ only by `State` generic of `SyntaxTree`, which is in fact unused. Converting both of the callsites to a version where `State` is `()` and implementing a helper to drop state reduces the API to single non-generic function.

Reviewed By: shiqicao

Differential Revision: D19443982

fbshipit-source-id: 3813e54d4c5597169a3208c49aeb4fa4c61b13be