Split rust_aast_parser_types from aast_parser
Summary:
`aast_parser` is supposed to be a simple glue crate coordinating between parser, lowerer and error checker. It takes surprisingly long to compile, and I traced one of the reasons to `rust_aast_parser_types`, so splitting it off to reduce cost of critical path.

I suspect the reason it takes so long is following definition:

```
#[derive(Clone, Debug, OcamlRep)]
pub struct Result {
   ...
   pub aast: std::result::Result<aast::Program<pos::Pos, (), (), ()>, String>,
   ...
}
```

`aast::Program` is a big generic, and this type definition requires instantiating some of its methods (i.e. to derive `OcamlRep`, or even `Drop`).

Reviewed By: shiqicao

Differential Revision: D19458719

fbshipit-source-id: 51382759373edc74812137431fc83fc4a0bd82e3