Fix bad position bug
Summary:
For legacy lambda inference, the position of a typing error was derived from the wrong type, and so the "primary position in a different file" assert would fire.

Fix is easy, but I've removed `new-inference-lambda` as the default for tests, putting it in `HH_FLAGS` instead. We had no regression tests for new-inference-lambda: now we do, in `test/typecheck/lambda/lambda_legacy`.

Reviewed By: manzyuk

Differential Revision: D19553634

fbshipit-source-id: c05df0a8916deca4254f9e3a3295262e7429d28f