don't recreate global type var if it already exists
Summary:
When localizing a global type variable, the type variable might have already been localized and therefore be already in the environment, for example when typechecking a method: we localize once while checking that the class does extend its parent, and once again while typechecking the body of the method.
Therefore, we should make sure we do not erase and recreate a type variable that already exists.

Differential Revision: D19539000

fbshipit-source-id: c0c58f60506866fd00b7263e72893a5a57bd78b3