Remove bizzare workaround from ext_icu_calendar.
Summary: IntlCalendar::set and IntlGregorianCalendar::__construct were packaging their arguments up into arrays, and calling private native implementations that took a single array arg instead of individual arguments. Comments indicate that this was because the "ActRec API" was not yet available and (presumably) having 6 arguments was not supported in whatever preceded it. This restriction has not existed for years, so stop doing the silly argument shuffling and just pass them normally.

Reviewed By: markw65

Differential Revision: D19559160

fbshipit-source-id: b2a37bcd6b77313365412fcc3de124e6f1b94622