Make the caller choose whether to open the database in-memory or not
Summary: I don't think there are many use cases where the fallback-to-empty-in-memory-database behavior of `Names::new` is desirable, and it could even be dangerous if it kicks in in an unexpected way. I think the caller knows which behavior they want, so we should ask them to choose by providing two constructors.

Reviewed By: losvald

Differential Revision: D19523522

fbshipit-source-id: efcdebe39eeb52a8becafdb85e1907fd5fbd065a