raise exception when setting variance on solved type var
Summary: Previous behavior was: when calling set_tyvar_appears_covariantly on a var, if the var didn't exist, spring the var into existence and set variance, if the var was already bound, sneakily discard the binding and map it to an empty set of constraints instead. Since we don't want these kinds of silent faulty behaviors, we raise an exception instead in those two cases. Previous diffs in this stack make sure those exceptions are gracefully surfaced as errors requiring to hit the nuclide bug button.

Differential Revision: D19557925

fbshipit-source-id: a0adcc279d5b4876a4cb252c43eb2924f459b108