fix of type variable being discarded with env
Summary: When subtyping Tdestructure, in some cases we create a Traversable<#1> where #1 is a fresh type variable (see inline comment for where this happens). This Traversable type ends up in a proposition which is sometimes ends up being conjuncted with an unsat prop created with `invalid_with` (instead of `invalid_env_with), which discards the environment containing #1. To fix that, we replace the calls to `invalid_with` with `invalid_env_with`.

Differential Revision: D19534215

fbshipit-source-id: b2da7af8e403bac705a74b4ddcf60891756dade8