Serialize RelativePath as a string
Summary:
The Errors type contains a map from RelativePath to the errors in that file. Unfortunately, this can't be serialized by serde_json (using the current, derived serialization we have for RelativePath). Backends for serde are free to refuse (at runtime) to serialize data structures that don't match the serialization format's expectations, serde_json expects maps to be serializable as JSON objects, and JSON objects can only have strings as keys, so **only data structures which serialize into strings can be used as map keys in serde_json**.

This is fine, I suppose--we just have to serialize RelativePath as a string instead of an object. This diff replaces the derived Serialize/Deserialize implementations with one based on the behavior of `storage_to_string` and `storage_of_string` in relative_path.ml.

Reviewed By: losvald

Differential Revision: D19458250

fbshipit-source-id: 4a5fa0acb154ef083b5536ef6390384dc319e685