Use Memcache to cache facts
Summary:
This adds a new MemcacheDataType to fetch and store Facts. Note that the cache isn't actually *enabled* until the associated configerator diff (D19340827) lands.

I'd like to make this available to `HH\facts_parse` as well so SQLFacts can benefit, but it's nontrivial since Memcache and DataType are closed-source. We can create a closed-source alternative builtin which uses caching, or we can hide the closed-source code inside a pImpl. Either way, I'd rather do that in a separate diff.

Reviewed By: dneiter

Differential Revision: D19343131

fbshipit-source-id: 6678fe1a93a31d85031be00358ebca8858fc0f45