Speed up the integration test by type checking all expected files in a batch
Summary:
The integration test (`//hphp/hack/test/integration:extract_standalone_test`) runs `hh --extract-standalone` on each supplied function name and makes two assertions: that the output of extraction matches some expected value and that it is well typed.  In particular, `hh_single_type_check` is run separately on the output of extraction of each function, and this is what makes this integration test slow.

To speed it up, I'm proposing to restructure it slightly: first run the extraction for each function name and confirm that the results match the expected values, and then type check all the expected files in a batch, using `--batch-files` option of `hh_single_type_check`.  (This is what makes the type checker test suite relatively fast.)

Reviewed By: CatherineGasnier

Differential Revision: D19581471

fbshipit-source-id: 03fda9b1b5333969f18992a2341727abdf797ba5