Change impl of Display for RelativePath to match impl of Serialize
Summary:
In {D19458250}, the implementation of Serialize for RelativePath was changed to match the behavior of `storage_to_string` and `storage_of_string` in relative_path.ml.

The implementations of Display for Prefix and RelativePath produce results similar to the implementation of Serialize, but slightly different (we print "dummy" instead of "", and use a space character to separate the prefix and path instead of a pipe). I think it'd be less surprising if Display behaved the same way as Serialize (and `Relative_path.storage_to_string`) does.

Reviewed By: shiqicao

Differential Revision: D19583413

fbshipit-source-id: d4697c33a9e8871f4177df860edaa15e54b77276