what saved-states are already there?
Summary:
I had made a change in OnDemandWWW to prefetch the appropriate saved-state.

But what I'm seeing in the logs is that only 20% of time time that a user starts ServerlessIDE on OnDemandWWW do they benefit from an already-downloaded saved state.

I don't have any idea as to why. But my plan with this diff is as follows. Any time saved-state-init is forced to download a saved-state, let's record all the saved-states that are on disk, so we get an idea of whether something had already been downloaded, and if so then what.

Reviewed By: arxanas

Differential Revision: D19570392

fbshipit-source-id: cf16b366ad08fa1cd13e350074b6773f430b703c