Fix assert when doing hhbbc dce on BaseSC
Summary:
When doing "minstr_touch", we assert that there hasn't already been an
action registered for the current bytecode for all the affected stack
slots. This is true in general, but not for BaseSC. BaseSC has two
immediates, so we need to do minstr_touch twice. The first one may
have registered an action at a stack slot, so the second one might
assert. Instead of asserting that no action has been registered,
assert that (if it exists) it is a MinstrStackFixup, and modify the
bitset appropriately.

Reviewed By: markw65

Differential Revision: D19606252

fbshipit-source-id: 8bac2f9a68701319632c9b7b12c39aa37014bb22