Set vmpc properly when resuming to throw an exception
Summary:
When we resume a suspended async function in order to throw an
exception (because the wait-handle failed), we set the vmpc to the
resume offset. However, the resume offset is the next bytecode after
the Await. This means when we go to look up a catch block, we'll look
up the catch block associated with the next bytecode and not the
Await. We seem to get away with this now, but this started to become a
problem when I experimented with some optimizations. Instead set vmpc
to point at the suspend offset (which is the Await/AwaitAll bytecode)
when we're going to throw.

Reviewed By: markw65

Differential Revision: D19605242

fbshipit-source-id: 89eca31db176bce2f6d044036832ede08e13babe