Fix server crash
Summary:
Consider the following example:
```
<?hh
// Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.

abstract class A {
  public function f(): void {}
}

final class B extends A {
  use C;
}

trait C {
  final public function g(): void {
    $b = $this->asB();
    $b->f();
  }

  private function asB(): B {
    return $this as B;
  }
}
```
`hh --extract-standalone "\\C::g"` currently crashes the server and causes it to restart indefinitely.  It turns out that the crash is causes by an infinite loop (and the stack overflow it entails), and the infinite loop occurs because we install the `do_add_dep` dependency callback before we try to extract the name of the file in which the target function is defined, and the latter involves fetching a function or class decl.  So, we try to fetch that decl, that might trigger a call to `Typing_deps.add_idep`, which executes all the installed callbacks, including `do_add_dep`, which then calls `is_builtin_dep`, which requires fetching the filename of a given dependency, which requires fetching a suitable decl, and so on.  With a clever inheritance hierarchy (as above) we enter an infinite loop.  Running the type checker (with just `hh`) first and letting it finish before running the extraction command would avoid this infinite loop because, I suppose, the decls would be fetched and cached already.

Once this infinite loop is fixed (by fetching the filename of the target before type checking it with dependency tracking turned on), the above call to `hh --extract-standalone` produces an error `Dependency not found: Method C::f`.  It turns out that when fetching a class element that represents a given class member by its name we unconditionally add a dependency on the member, regardless of whether it actually exists in the class or not.  After fixing that the above extraction goes through successfully.

Reviewed By: Wilfred

Differential Revision: D19601825

fbshipit-source-id: 6857a509488051a5abe9b0a8c56db36380ae992f