Add BVecLike and BVArrLike types to hhbbc
Summary:
For TFunc we introduced TStrLike to encode the fact that in many contexts TFunc would behave like TStr. For TClsMeth we need TVArrLike (behaves like a TVArr), TPArrLike (behaves like a TArr, a supertype of TVArrLike for Array type tests), and TVecLike (behaves like a TVec for when the DVArr flag is thrown).

We augment `loosen_string_like` to also loosen TClsMeth types which fixes many of the same comparison bugs fixed for TFunc in D19565006

Reviewed By: ricklavoie

Differential Revision: D19583306

fbshipit-source-id: 4c981b3e5b7a2bb9b6e000536de4bfe4654f11ec