Fix ProvenanceSkipFrame handling of collections
Summary:
We have two bugs around collections literals in ProvenanceSkipFrame (PSF) functions. First of all, when we emit code, we try to tag the collection with array provenance information rather than the original array. We fix this in the emitter in hackc.

This bug revealed a more serious refcounting bug in HH\tag_provenance_here: for inputs that don't need provenance, this function doesn't touch the input value, meaning that it gets destroyed when the native call returns. We fix this code by using tvReturn rather than returning the raw value. tvReturn is equivalent to inc-ref-ing it and then converting uninit to null; the latter effect does nothing for input parameters which are already guaranteed initialized. We could also make the inc-ref explicit, but we don't want to do so in extension code.

I wrote a unit test that exercises these new code paths. "test_collection_literal" exposes the common case where the tag is implicitly done by the PSF annotation; for this case, we check that the inner Hack array has provenance information and the outer collection does not. However, this test case is fixed by the emitter changes and no longer hits the ref-counting bug. As a regression test for the refcounting but, we explicitly call HH\tag_provenance_here on a collection; without the extension code fix, this case would still segfault.

Reviewed By: jjgriego

Differential Revision: D19626295

fbshipit-source-id: 13cc408505fb047d6564ac4b893bc7a3b9d5db00