Support missing type hints
Summary:
Missing type hints (e.g., missing function parameter or return hints, missing generic type parameters) are represented internally as type hint `Hany` (only in partial mode files; in strict files missing hints are usually a type error, but can be suppressed with HH_FIXMEs...).

At the moment, the pretty printer of hints fails on `Hany` causing the extraction tool to crash on any function that has a dependency containing a missing type hint.  This is suboptimal and I'm proposing to improve this by generating a special type `EXTRACT_STANDALONE_ANY` that behaves like the "any" type.  (It is possible to defined such a type using an HH_FIXME; basically, this is a hack that abuses the HH_FIXME mechanism to create a type that internally is represented as `Terr`, which for all purposes behaves identical to `Tany`.)

Reviewed By: Wilfred

Differential Revision: D19622512

fbshipit-source-id: 436c4cc2bd743e3e58e48b957b06e8f4e918dc21