don't connect until hh_server is ready
Summary:
A number of users are getting "unexpected command needing writes in priority channel: Cleaning up persistent client"

The only codepath that can lead to this is (1) there's an existing LSP connection to hh_server, (2) hh_server is currently performing a length operation, (3) a new LSP connection is made.

Explanation: hh_server can only support a single LSP connection at a time. If a new one arrives, then it has to boot the old one off. Booting off constitutes a "write" operation since it might alter hh_server's state.

In D7923582 in May 2018, we'd made it so that a new LSP connection request would go on the "high priority channel", i.e. it would be able to be performed even if a lengthy opration is in progress.

But what happens in these bug reports is that the new LSP connection involves booting off the old IDE connection, which involves mutating state, which is disallowed during a length operation. And so it crashes.

This diff makes it so that "connect new LSP client" no longer uses the priority channel. It will instead wait until hh_server is able to handle writes.

This will have the downside that some connections will take longer. Under ServerlessIDE, users won't get errors for longer until hh_server is ready. For users not yet on the ServerlessIDE beta, they won't even get hover or GoToDef or anything working until hh_server is ready. (but at least the status indicator will correctly show them that we're waiting for hh_server to settle).

Reviewed By: arxanas

Differential Revision: D19567413

fbshipit-source-id: 21ccb749282f03c5b67005e786750115a23ea2b5