eager methods return list not Sequence.t 2/2
Summary:
There are several functions in Typing_classes_heap which are declared as returning a Sequence.t, even though their internal implementation is wholly eager.

This diff changes them to return a list, more truthfully reflecting what they do.

It might be considered good to return Sequence.t so that, at some time in the future, we could change their implementations to truly be lazy, and all the callers would hopefully benefit from it.

On the other hand, we don't have the right telemetry in place to know where to gain lazinesss benefits. And adding that telemetry is actually made easier by having them not return sequences. Also, once we move to a decl service, then having a chatty "Sequence.t" API will be harmful.

So I'm simplifying things down and using just a "list" where appropriate.

This is diff 2 of 2, both doing the same thing.

The important place to review is Typing_lazy_heap.mli and Typing_lazy_heap.ml. All the other numerous file-changes are just trivial consequences of these two.

Differential Revision: D19639093

fbshipit-source-id: f8f4c67f1f7d1be32c8b9e1a98477c10d4b4b6d9