Disallow discarded nullable awaitables improvements
Summary:
I tried to roll out the new `disallow_discarded_nullable_awaitables` flag, but found that many of the errors were bogus due to us not including `??` in the allowed cases.

Also I found the previous logic quite difficult to follow, so I rewrote it to make more sense to me. Would love feedback on the changes.

Differential Revision: D19606008

fbshipit-source-id: 70821e906c5c14450f15fc462f8ad8cfa8abc4a9