counters module
Summary:
My goal is to keep track of how many decl cache misses there are, and how much time we spend on them.

This is a new module, intended to be used for keeping track of count+time. I will use it in compute_tast and also process_file.

The functionality in this file used to be in Deferred_decl.ml, but Waleed made the observation that it didn't really belong there. So I'm moving it out and doing it cleaner.

Reviewed By: arxanas

Differential Revision: D19639122

fbshipit-source-id: 921e429a8809c5e0c7ff1c84bf0194ccd9175ac6