allow tests to inspect telemetry
Summary:
I think it'll be useful for tests to be able to inspect Telemetry.t objects.

For instance, server_test can inspect how many decls had to be fetched from the cache, or how many deferments happened. These are numbers that it makes sense to record in our telemetry reports, but it doesn't make sense to expose in a strongly-typed fashion all the internal details about them to the bulk of the hack codebase, but it does make sense for unit tests to dig into these kinds of telemetry counters.

None of the json accessors were particularly well suited to this purpose. For tests, we want json accessors that ensure that a field is present. So I wrote some handy accessors.

Reviewed By: arxanas

Differential Revision: D19669442

fbshipit-source-id: 15855dfa6037a830a88a5b08b29204ba967d0549