Include key as parameter to `Lru_cache.get_size`
Summary: It turns out you can't examine the `value` unless you match on the `key` first, because it's a GADT. That means that it's currently not possible to actually look at the value for `value` when deciding on the `size` of it. The only meaningful implementations are constant values like `1` or magic functions like calling `Obj.reachable_words`.

Reviewed By: jewelpit

Differential Revision: D19631931

fbshipit-source-id: d878d6df4a913b41c19cd3b4d17c665df431e6f5