Fix refcounting bugs in opt_tag_provenance_here
Summary: I forgot to inc-ref the value returned by this function. This function also didn't handle the NativeImpl version of the ABI (which we don't care to optimize, since it's only used in non-repo mode). These bugs were asymptomatic because we currently only use tag_provenance_here with static arrays, but we want to change that. I added a new unit test that hits both issues (the first, in repo mode, and the second, in non-repo mode).

Reviewed By: viratyosin

Differential Revision: D19715929

fbshipit-source-id: c5b3f2e4977c959f506bf88818deb4806f447d40