Delete dead, broken iteration functionality.
Summary:
IterateV/IterateKV allow passing a bool as PreCollFn (in order to bail early if you have anything other than an array), and there is a 3-arg variant of both that passes false for PreCollFn to the full version. Luckily this flavor is dead, because a bug has crept in: in this variation, we short-circuit before checking for ClsMeth, which means ClsMeth is not treated the same as an array, which would be a bug.

Delete the unused, buggy short-circuiting and the 3-arg flavor of IterateV/IterateKV that accessed it.

Reviewed By: ricklavoie

Differential Revision: D19714317

fbshipit-source-id: efbf604146cd95ab5c5f7e1b23bd6e12b0c83e77