Setting for loading the naming table on full init
Summary:
Loading the naming table on full init should be independently configurable. I want to be able to not load the naming table on full init if remote type checking is enabled.

CONTEXT: Normally when we do a full-init, we spend a few minutes updating the naming table before we start doing typechecking.

It would be a nice optimization (in the case of distributed typechecker) to avoid those initial few minutes, so that the entirety of distributed-typechecking can start immediately. That's what the code used to do prior to this diff - it allowed a full init but based on a naming-table saved state.

There were some issues. If there were duplicate names in the SS, then the naming-table-saved-state didn't even include them. There was also T61479318. (and actually, I can't see names being updated correctly? maybe they are?)

So this diff is a way to disable that optimization, to avoid having to deal with those issues just now.

Reviewed By: ljw1004

Differential Revision: D19646489

fbshipit-source-id: 75fde64972c18d23ed8a029a92549185682fe01b