Rewrite prototype and work-orchestrator in Rust
Summary: I'd like to start looking into having the work-orchestrator distribute work to many worker processes in parallel. I believe this would be easier to do in Rust, so this diff rewrites the work-orchestration logic in a new Rust binary, hh_scheduler. This would make it difficult to keep the marshalled-OCaml-blobs IPC format for the orchestrator-worker IPC protocol, so this diff uses a new protocol which exchanges newline-delimited JSON messages. To do so, this diff adds a Rust layer around the OCaml worker logic. It also takes this opportunity to rewrite the worker-prototype logic in this Rust layer (hopefully reducing the need for duplicated logic between languages for tasks like deriving a socket file name from a root path).

Reviewed By: ljw1004

Differential Revision: D19418397

fbshipit-source-id: 218c4b9fba7bebaa71f5cb6c2413bc6228fcde7b