Report parsing errors
Summary:
It turns out that `compute_tast_and_errors` does not report parsing errors. They're actually dropped unless we pass the `keep_errors` flag to the parser.

This commit changes it so that we do collect the parsing errors, and merge them with the typing errors when it's time to compute errors.

This suggests that there should be a mode of operation that skips producing parsing errors if they're not necessary. I'll leave that for another time, if it turns out we need to rely on that.

Reviewed By: 2BitSalute

Differential Revision: D19277781

fbshipit-source-id: dd2cfb18b12fe4788961165e57b9c25f2856d25d