Look up symbol paths one at a time
Summary: Currently, our lookup APIs look up symbols in batches, possibly to avoid the overhead of preparing the statement many times. However, this is a bit awkward for most use cases, where we have a single symbol in hand and want to learn the file in which it was defined. Instead of providing a bulk API, we can use rusqlite's built-in prepared statement cache and provide an API for looking up a single symbol at a time.

Reviewed By: Wilfred

Differential Revision: D19587808

fbshipit-source-id: d10a1cdcfbbac20e41164c2c2d5f1287c0268a21