Move table-specific logic to toplevel functions
Summary: In {D19523025}, I removed the table-specific structs in order to do away with the Arc<Mutex> they used to share the Connection. Instead, I put query methods directly on the Names struct, and split out its implementation into table-specific submodules. This works okay, but it might be a bit clearer to put all of the Names methods in one place and to hide implementation details of Names from the table submodules. This diff does so by moving all the table-specific methods to be toplevel functions within the same submodule, and re-adding forwarding methods (e.g. paths_of_consts) to Names (as we had with the table structs).

Reviewed By: Wilfred

Differential Revision: D19583627

fbshipit-source-id: 3632ce4d585508f9af368d42e239fe860329a6c1