Add implicit coercion for class_meth() on property type hints
Summary: When `fun()` and `class_meth()` were first added we were still rolling out property type hint support and the decision was made to not coerce for these types. While this has been fine for `fun()` we are now fully enforcing property types everywhere and it's not clear there's a safe way to move forward with `class_meth()` without a migration path.

Reviewed By: ricklavoie

Differential Revision: D19706007

fbshipit-source-id: f103b7af17227bc5160f661e7cc806e5c6442621