Ban construction of class_meth() with Closure::__invoke
Summary: If a short lambda does not otherwise need to capture `$this` it can be marked as `static` to avoid extending the lifetime of the object constructing it. The method itself still requires a `$this` in its prologue (namely the Closure instance) and therefore cannot be invoked via a `class_meth`.

Reviewed By: markw65

Differential Revision: D19708002

fbshipit-source-id: 6753a0ac11ef5594e4940113351f67b8aa2dcd5b