telemetry for number of results
Summary:
There are anecdotes of definition requests being handled but giving 0 results. We're not currently tracking this - our telemetry counts any response as success, even a response with 0 results.

This diff adds a general "result_count" field to the telemetry for each LSP request. It records how many answers it contained. I have these uses in mind:
1. If we do an A/B test and observe a marked increase in the number of things giving "0" results for B but not A, then something's wrong
2. If the user goes to the trouble of doing cmd+click on a symbol, it's hard to imagine them doing that accidentally, so every single definition request should have >0 results. (Well, there are exceptions like accidentally hitting the keystroke for it, but these exceptions are probably fairly rare).
3. If we in future augment the completion request with an indication of whether the completion request came right after `->`, then the user expects every one of these to have >1 result

It doesn't make sense to compare (say) result_count of autocomplete with that of hover. I just decided to add result_count universally across all result types in anticipation of some future needs, where we will find we need to look at something and are grateful that the telemetry was already in place.

I don't expect a single int column to add much size to our telemetry datastore. Am not even bothering to calculate it.

Reviewed By: arxanas

Differential Revision: D19686620

fbshipit-source-id: 79a92ca79f841e393d596c6659296ecd6aeb8021