Reduce the size of HHBBC::FCallArgs
Summary:
FCallArgs is a big struct, but most of the time we can get away with something much more compact.

This introduces FCallArgsSmall, which fits into 8 bytes, and covers almost all calls; and FCallArgsLong which is basically the original FCallArgs.

FCallArgs itself becomes a union of FCallArgsSmall, and copy_ptr<FCallArgsLong>, using the top bit to differentiate.

This drops sizeof(Bytecode) from 40 bytes to 32. I'll go after the next biggest offenders in follow up diffs.

Reviewed By: ricklavoie

Differential Revision: D19750603

fbshipit-source-id: 9feec782ab0637c890723402caf5081d454a22f0