Explicitly mark bytecodes using local names, or local indexes
Summary:
This is the first diff working towards enabling compaction of local slots.  By adding a name index to bytecodes that may need the name of the local being operated on, we are free to remap the local slot index for the purpose of sharing slots.

This adds the bytecode immediate types ILA, and NLA.  It alters the meaning of LA
- ILA is a local operand that must know the slot index
- NLA is a local operand that must know the name of the local being operated on
- LA is a local that must have a pointer to the local slot

This also fixes up an issue with DCE done for minstr final operations.  This issue was uncovered while reducing CGetL to CGetQuietL.

Reviewed By: ricklavoie

Differential Revision: D19730560

fbshipit-source-id: 73e35add97e5c7f4dc20bbd28d9c2750680c1751