Fix possible aliasing issue in VregSet
Summary:
One of the VregSet tests started failing when I tried to switch to
gcc-9, and looking at the code, it seems to be related to aliasing
between extended.bit and blocks.data[1].

I think that code isn't strictly correct, because blocks and extended
don't (strictly) have a common initial sequence, even though there are
static asserts ensuring that the layout *is* the same.

I initially changed bit to an array of unsigned char, and indexed into
that array; that fixed the issue, but I realized that strictly its
*still* not ok (without the common initial sequence, accessing bits is
still ub).

So I just went with casting the union to uint8_t* and indexing into
that array (which also works).

Reviewed By: ricklavoie

Differential Revision: D19755978

fbshipit-source-id: e6728ff62ef950c93b396fd06aaab48b02847c05