Remove AutoloadMap::getTypeAliasFile
Summary:
This is a companion diff to D19727144, but neither diff is dependent on the other. See that diff for the justification.

I'm starting the merging process by removing `AutoloadMap::getTypeAliasFile()`. I've made `getTypeAliasFile()` a private method for each of `AutoloadMap`'s callers, which each implementation of `AutoloadMap` uses internally for now.

Reviewed By: billf

Differential Revision: D19732231

fbshipit-source-id: eabaf09efe1cc4b4d7f38fcc7a347a63b74435b0