Test new function call syntax
Summary:
Previously, we did not allow call expressions with an empty targ list. i.e. `foobar<>($args)` was not allowed.

With the changes for function pointer and the new syntax for function pointers, this will be allowed so as to be consistent with the new function pointer syntax. Thus, `foobar<>($args)` will now be legal.

Of note:

`$obj?->foobar<>(expr)` will parse as a nullsafe call such that `expr` is evaluated regardless of `$obj`'s null-ness whereas `($obj?->foobar<>)(expr)` will parse as a nullsafe access of the method and will be a type error.

Reviewed By: Wilfred

Differential Revision: D19604294

fbshipit-source-id: 4ae1172aa49e075cd2b0fab266169d0e067f767b