Fix issue with jumps to fallthru Vinstrs
Summary:
When emitting jumps to a block which contains nothing but a fallthru
instruction, the relocator will not update the jump target
appropriately. This is because no machine code is actually emitted for
a fallthru, so the jump ends up jumping to the memory address after
the block. The relocator sees this as a jump to outside of the
relocated region, so does not touch it. The fix for this is
simple. Emit a nop instead of nothing for a fallthru. The relocator
will remove the nop, and it makes the jump target lay inside the
relocated region.

Reviewed By: oulgen

Differential Revision: D19793352

fbshipit-source-id: 01dc8c2372541d4b6b9b391abfcdbb0e41154562