Add support for per file code coverage
Summary: When enabled in sandbox mode introduce guards in the entry translations for functions to check if code coverage for the file defining the function is enabled. Coverage itself is still collected in the interpreter and stored using a dynamic_bitset in rds (for this mode of code coverage we only care about lines covered and not frequency).

Reviewed By: oulgen

Differential Revision: D19757824

fbshipit-source-id: 9b3047630163582a6a811a960d4a918a66e2d31b