use Lsp.Error.t more
Summary:
I want to surface external errors better to the user, e.g. "SSL certs are bad"

To that end, I want the Hack output window in VSCode to be relevant and actionable.

To that end, I want to eliminate the spam that we see in it. The biggest spam is all the error messages that show up when the server is stopped or unavailable. VSCode and Atom behavior is that all errors other than RequestCancelled show up in the output window. Currently, all errors reported by ClientIdeService.ml are surfaced as UnknownErrorCode errors.

To that end, I want to push the choice of LSP error code further down the stack. I think the best abstraction is for ClientIdeService.rpc to be able to return a fully-specified Lsp.Error.t in case of error, rather than just returning a Marshal_tools.remote_exception_data and trusting clientLsp to convert it into the right Lsp.Error.t.

That's what this diff does. It makes ClientIdeService.rpc return an Lsp.Error.t rather than remote_exception_data.

This change has brought additional changes with it. It used to be that the only thing we returned was "message" and "stack". Now instead we're returning "code, message, and JSON data". In other words, (1) the error-generation-sites have more flexibility in what kind of data to produce, and are not limited solely to data; (2) they are responsible for encoding whatever data they want, including callstacks, in that data.

This also means I'm changing what gets logged to telemetry. Prior to this diff, we always had two things, "message+stack". Now we have "message+data". So I'm logging that data instead.

It used to be that when ClientIdeService returned a remote_exception_data, then we bundled that up in an exception, and later on we deconstructed that exception. This diff changes ClientLsp.report_error to avoid that. Instead, now, we keep it intact as an Lsp.Error.t.

It was a bit goofy how we handled stacks before. If clientLsp had issued a request to VSCode, and got back an error response, then our "Lsp.ErrorResult" object had a reasson and a stack, and we had to invent an (empty) stack. In other words, our use of stack wasn't very LSP-ish, not round-trippable. Now, thanks to this diff, the ErrorResult object solely stores an Lsp.Error.t and it's clean.

Reviewed By: arxanas

Differential Revision: D19800848

fbshipit-source-id: 80b218f9f191ab190fd3717ca868c313cdca3d66